fix: Run tests inside Docker container instead of on Pi host
The integration-test job was trying to run 'go test' on the Pi host,
but Go isn't installed there (and shouldn't be - OpenCV setup is complex).

Changed to run tests inside the running gizmatron container using
'docker exec' commands. The container already has:
- Go 1.23.5
- OpenCV with contrib modules (ArUco, etc.)
- All project dependencies

This allows tests to run with full environment without needing to
install Go/OpenCV on the Pi host.